Move dto package out of controller package
